chore: moved fetch colours logic to context
