+++ src/contexts/ColourContext.tsx
@@ -1,9 +1,31 @@
-import { createContext, useState, ReactNode, useContext } from "react";
+// ColoursContext.tsx
 
-const ColoursContext = createContext<{ colours: { hex: string }[]; setColours: (colours: { hex: string }[]) => void } | undefined>(undefined);
+import { createContext, useState, ReactNode, useContext, useEffect } from "react";
+import { fetchData } from "../utils/fetchData";  // Update this import path
+
+interface Colour {
+  id: number;
+  hex: string;
+}
+
+const ColoursContext = createContext<{ colours: Colour[]; setColours: (colours: Colour[]) => void } | undefined>(undefined);
 
 export const ColoursProvider = ({ children }: { children: ReactNode }) => {
-  const [colours, setColours] = useState<{ hex: string }[]>([]);
+  const [colours, setColours] = useState<Colour[]>([]);
+
+  // Fetch colours from an API when the provider is mounted
+  const fetchColours = async () => {
+    try {
+      const data: Colour[] = await fetchData("colours", "GET");
+      setColours(data);
+    } catch (error) {
+      console.error("Error fetching colours:", error);
+    }
+  };
+
+  useEffect(() => {
+    fetchColours(); // Fetch colours when the provider is mounted
+  }, []);
 
   return (
     <ColoursContext.Provider value={{ colours, setColours }}>
@@ -12,6 +34,7 @@ export const ColoursProvider = ({ children }: { children: ReactNode }) => {
   );
 };
 
+// Named export for the hook
 export const useColours = () => {
   const context = useContext(ColoursContext);
   if (!context) {
